Split fxrecord into a workspace
This will enable us to split fxrunner and fxrecorder into bins and libs
(without sharing all of the implementation in both of them), allowing
for integration testing between the two.
