+++ fxrecorder/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "fxrecorder"
+version = "0.1.0"
+authors = ["Barret Rennie <barret@mozilla.com>"]
+edition = "2018"
+license = "MPL-2.0"
+
+[[bin]]
+name = "fxrecorder"
+path = "src/main.rs"
+
+[dependencies]
+derive_more = "0.15.0"
+futures-preview = { version = "0.3.0-alpha.19", features = ["async-await", "std"] }
+serde = { version = "1.0.100", features = ["derive"] }
+slog = "2.5.2"
+slog-async = "2.3.0"
+slog-term = "2.4.1"
+structopt = "0.3.1"
+tokio = "0.2.0-alpha.6"
+tokio-serde-json = "0.3.0"
+toml = "0.5.3"
+
+[dependencies.libfxrecord]
+path = "../libfxrecord"
