+++ fxrunner/src/main.rs
@@ -8,14 +8,14 @@ mod proto;
 use std::error::Error;
 use std::path::{Path, PathBuf};
 
+use libfxrecord::{run, CommonOptions};
 use slog::{info, Logger};
 use structopt::StructOpt;
 
 use crate::config::Config;
-use fxrecord::{run, CommonOptions};
 
 #[derive(Debug, StructOpt)]
-#[structopt(name = "fxrecorder", about = "Start FxRecorder")]
+#[structopt(name = "fxrunner", about = "Start FxRunner")]
 struct Options {
     /// The configuration file to use.
     #[structopt(long = "config", default_value = "fxrecord.toml")]
@@ -29,19 +29,20 @@ impl CommonOptions for Options {
 }
 
 fn main() {
-    run::<Options, Config, _, _>(fxrecorder, "fxrecorder");
+    run::<Options, Config, _, _>(fxrunner, "fxrunner");
 }
 
-async fn fxrecorder(log: Logger, _options: Options, config: Config) -> Result<(), Box<dyn Error>> {
-    use crate::proto::RecorderProto;
-    use tokio::net::TcpStream;
+async fn fxrunner(log: Logger, _options: Options, config: Config) -> Result<(), Box<dyn Error>> {
+    use crate::proto::RunnerProto;
+    use tokio::net::TcpListener;
 
-    let stream = TcpStream::connect(&config.host).await?;
-    info!(log, "Connected"; "peer" => config.host);
+    let mut listener = TcpListener::bind(&config.host).await?;
+    let (stream, addr) = listener.accept().await?;
 
-    let mut proto = RecorderProto::new(log, stream);
+    info!(log, "Received connection"; "peer" => addr);
+    let mut proto = RunnerProto::new(log, stream);
 
-    proto.handshake().await?;
+    proto.handshake_reply().await?;
 
     Ok(())
 }
