+++ fxrunner/src/proto.rs
@@ -2,10 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+use libfxrecord::net::*;
 use slog::{info, Logger};
 use tokio::net::TcpStream;
-
-use fxrecord::net::*;
 
 /// The runner side of the protocol.
 pub struct RunnerProto {
